Fixed Camera
Planning on making camera cooler..but the scrolling at least works now
without the player going out of view. We need to figure out how to keep
the player from going out of the boundaries. I tried but didn't have any
luck so I reverted all of that back to what it was before I touched it.
I believe that needs to be done in the player class based off of the
player's position. The camera can be modified after that is figured out.