Added changes to animation, player, maptile
MAPTILE CHANGES:
Added a new boolean for unstable blocks, which still needs to be filled
in from the XMLs. Added options for resetting the animations of certain
tiles.

PLAYER CHANGES:
Added some tile method calling when player hits a breakable or unstable
block.

ANIMATION:
Added a boolean at the end for the animation to loop repeatedly on the
first frame -- used for blocks that need to preserve their first frame
until interacted with.

TODOS:
Need to change how tiles are added in World to account for the Unstable
Blocks. Need to create some basic animations for blocks itself, and
player death. Also, change functionality of the game so that if player
dies, all tiles are reset and player animation for death plays out, then
resets to start.

Also added some placeholder blocks.