Added code to reset level when player dies.
Bugs Found:
+ Death animation sequence doesn't occur.
Issue: Game updates the world class first before player / animations
class.
+ Collisions
Issue: TBD -- still needs debugging. I will be making changes to the XML
later to have all the trap tiles be in front of the MapTile[], but this
shouldn't make a huge difference on why this is occurring on the scale
that the player can walk on 3/4th of one trap block before dying.

Solutions I've tried for the first bug:
+ Made Animations class public and deathAnimation in Player class public
to add and access a boolean that checks to see if the animation has been
played out. [This has failed but I may have placed it in the wrong area.
Unlikely since it's probably due to the hierarchy.]
+ Created a separate method that changes the levels in world. Created a
world instance in player class and called that function. [Also failed,
probably due to the hierarchy.]

Solutions thought of but haven't tried for the first bug:
+ Create a method that switches the visibility of deathAnimation to
visible instead of relying on the Draw method to execute each animation.
This means all animations will already be drawn, but will only be
visible when the boolean is true so Update will handle the animations.
It won't be visible if false. Probably requires changes to Animations
class to have an isVisible boolean.
+ Create a specific method just for death animation sequences to play
out and then call it in World using player.playDeathSequence.
+ Pretty much moving the dependencies of having the death animation
sequence execute to using Update, not Draw. Death animation sequence
should already drawn but just not be viewed as visible or accessible
(ex: Initial position of the death animation can be found in some random
position off the game window screen and be updated to move in the
position the player is currently at when colliding with the trap
block/dying).